% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfred_revision_analysis.r
\name{alfred_revision_analysis}
\alias{alfred_revision_analysis}
\title{Get Data on Series Revisions from ALFRED}
\usage{
alfred_revision_analysis(series_id)
}
\arguments{
\item{series_id}{Character string (required).  This is the ALFRED series ID that will be returned.}
}
\value{
A tibble with observations for each date in the series, showing the initial, final, change, percent change, number of items, and number of unique values across all observations for that date.
}
\description{
This function pulls data for a specified series from ALFRED, the archival version of the St. Louis Federal Reserve Economic Database, FRED.  It retrieves the data, then summarizes that data with the initial observation, the final observation, and the change between the two.
}
\examples{
\dontrun{

# Retrieving data for Nevada Total Nonfarm employment, seasonally adjusted.
nv_jobs <- alfred_revision_analysis("NVNA")

# Revision analysis for Nevada's Unemployment Rate
nv_ur <- alfred_revision_analysis("NVUR")
}

}
