% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_if_new.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the User Cache Directory for the Package}
\usage{
get_cache_dir()
}
\value{
A character string giving the full path to the cache directory.
}
\description{
Returns the path to the user-specific cache directory used by the package.
No subfolders are created; all files go directly in this directory.
}
