% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_qcew_data.R
\name{get_latest_qcew_data}
\alias{get_latest_qcew_data}
\title{Get the Most Recent BLS QCEW Industry Data}
\usage{
get_latest_qcew_data()
}
\value{
A list containing:
\describe{
  \item{data}{A `data.table` containing the BLS data for the latest available quarter.}
  \item{check_quarter}{The quarter of the data returned.}
  \item{check_year}{The year of the data returned.}
}
}
\description{
This function checks recent quarters (by default, two previous quarters)
of BLS Quarterly Census of Employment and Wages (QCEW) data for industry code 10,
and downloads the most recent available CSV file. The function uses
[download_if_new()] to cache data locally and avoid unnecessary re-downloads.
}
