% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_recession.r
\name{get_min_year}
\alias{get_min_year}
\title{Get minimum year from a date column}
\usage{
get_min_year(date_column)
}
\arguments{
\item{date_column}{A vector of Date objects.}
}
\value{
Numeric. The year of the minimum date in the column.
}
\description{
A convenience function to extract the minimum year from a date column,
useful for determining the `start_year` parameter for `geom_recession()`.
}
\examples{
\dontrun{
# Get the earliest year from your data
min_year <- get_min_year(economic_data$date)

# Use it with geom_recession
economic_data \%>\%
  ggplot(aes(x = date, y = value)) +
  geom_line() +
  geom_recession(start_year = min_year)
}

}
\seealso{
\code{\link{geom_recession}}
}
