% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{Streamline connection to WID Database}
\usage{
run_query(query, ODBC = "WID_DB")
}
\arguments{
\item{query}{Character. The SQL query that you with to execute.}

\item{ODBC}{Character. The name of an ODBC connection on your local machine that the script will use.}
}
\value{
The results of the SQL query, run with as.is=TRUE.
}
\description{
Wrap the con/query/disconnect workflow to query a SQL database into a function that allows use of a single function.
}
\details{
The function uses RODBC to run a query (required input) against a named ODBC connection (default "WID_DB").
Alternative ODBC connection name may be provided with the ODBC argument.
}
\examples{
\dontrun{
ces_query <- run_query("SELECT * FROM dbo.ces")
}
}
